+++ lib/src/flutter_rounded_date_picker_widget.dart
@@ -97,6 +97,7 @@ typedef SelectableDayPredicate = bool Function(DateTime day);
 
 Future<DateTime> showRoundedDatePicker(
     {@required BuildContext context,
+    double height,
     DateTime initialDate,
     DateTime firstDate,
     DateTime lastDate,
@@ -170,6 +171,7 @@ Future<DateTime> showRoundedDatePicker(
           //
         },
         child: FlutterRoundedDatePickerDialog(
+          height: height,
           initialDate: initialDate,
           firstDate: firstDate,
           lastDate: lastDate,